requirements.txt
beautifulsoup4==4.6.0
certifi==2018.4.16
chardet==3.0.4
Django==2.0.4
django-crispy-forms==1.7.2
django-extensions==2.0.7
django-postgres-extensions==0.9.3
django-taggit==0.22.2
django-widget-tweaks==1.4.2
html5lib==1.0.1
html5print==0.1.2
idna==2.6
numpy==1.14.3
ply==3.4
psycopg2==2.7.4
python-dateutil==2.7.2
pytz==2018.4
requests==2.18.4
six==1.11.0
slimit==0.8.1
tinycss2==0.6.1
urllib3==1.22
uWSGI==2.0.17
webencodings==0.5.1
